refactor(core): migrate RiceLeafClassifier to BasePredictor interface
